chore: initial setup of open source code
**WHY**

Previously the code was hosted in a private repository, but as
the tool grew requests were received to go open source. As I
value transparency and validity above all else this seemed like
a good step to take.

**WHAT**

Moved the original source code over from the private repository.
Added documentation, licensing, default reviewer information and
a code of conduct.