docs: setup Github issue templates (#3)
**WHY**

To ease the barrier-of-entry for new contributors and provide
the developers with clearer defined issue we'll set up some
issue templates.

**WHAT**

Added bug and feature issue templates. Added the option to create
blank issues without template. Added a direct link to Discord in
case the contributor just wants to discuss the tool.