fix: tier-list image lookup case-sensitivity failure
**WHY**

Looking up the static images used in the tier list failed on
Linux OS as the images were looked up with all caps, but the
names of the images were in lowercase. Linux is case-sensitive.

**WHAT**

Converted pokemon names to lowercase before looking up their
images. Also added a drive-by npm install line in the backend
README.