feat(backend): added island domain and utils
**WHY**

The previous/current solution for islands relies on query
parameters and is a bit messy. Now that we're getting new
islands the lack of easy scalability is notable.

**WHAT**

Added an island domain interface instead. The Island interface
can be used by the API instead of the query parameters. Each
island will be responsible for their own data such as
preferred berries.

refs #62