feat(backend): added lapis to api island filters
**WHY**

To prepare for the frontend the API needs to be expanded with
the option to set Lapis Lakeside filter.

**WHAT**

Added lapis to the query parameters. There is ticket #62 to
replace current query boolean island filters with new island
interface. This has been done here in part for the optimal set
and meal ranking.

Refs #62