feat(optimalset): refresh UI and added Lapis Lakeside
**WHY**

We needed to add Lapis Lakeside as an option to the optimal
set tool. Optimal set was also in need of a UI refresh and the
island filters specifically were confusing.

**WHAT**

Replaced limit to level 50 setting with option for custom level.
Updated UI to use range sliders and toggle buttons like the
production calculator tool does.

refs #55
refs #58