feat(backend): added optimal flexible ranking
**WHY**

The previous (current) flexible ranking on the website is one of
the oldest features. It's functionality is very basic, it just
checks percentage of recipe covered and ranks based on that.

With more sofisticated algorithms now in place in Sleep API we
can create a better ranking that takes the current meta and
available Pokemon roster into account.

**WHAT**

Added an API for /api/optimal/meal/flexible which takes a request
body for production input and method POST. For the given input
we calculate the production of all available Pokémon. Then we run
every possible team composition in the game against every recipe.

Finally we rank all Pokemon from the optimal solutions and find
Pokemon that can flex multiple optimal solutions, especially
for different recipe types.

This will be replacing the current meal ranking that's in place.
The new ranking will only take recipes into account for each
Pokemon for which they are able to provide an optimal solution.
An optimal solution means they're part of a team composition
of the smallest possible size for given recipe. So if Absol
can do spore mushroom salad with team of size 3 we don't
count it since Abomasnow is more optimal, it can do it with
size 2.