feat(teamfinder): improved Optimal Set settings
**WHY**

The name of the tool "Optimal Set" was confusing and explained
more about the algorithm behind it than it did what the tool is
actually to be used for.

We were missing description of the tool.

We had no direct selection of subskills.

**WHAT**

Renamed Optimal Set to Team finder.
Added detailed description of the tool, with assumptions and some
information about the algorithm behind it.
Added selection of subskills with default to optimal.