WIP: refactor PokemonIngredientSet
In order to consume Event data in the production
calculator, the PokemonIngredientSet type is
replaced by the PokemonIngredientList type as an
input. This allows the production calculator to
modify a Pokemon due to an event before generating
its ingredient list.

This commit is not close to done, but I'm
uploading it so I can ask for help with the error
message.