fix(common): add previous evolution count
The goal is to remove the maxCarrySize value to
reduce the size of the Pokemon definitions, thus
reducing the chance of errors when adding new
Pokemon. maxCarrySize is always carrySize + 5 for
each previous evolution.

This commit adds the new data. A later commit will
remove the old data, so that tests can be written
with access to both sets of data to ensure they
match.