fix(pokemon): remove maxCarrySize
Now that previousEvolutions is in the Pokemon
data, we no longer need to specify a maxCarrySize
for each Pokemon in the data, as we can compute
that from other fields.

A few other data types also had a maxCarrySize
field. I replaced those with inventoryLimit, in
large part to make it easier for me to search for
remaining instances of the Pokemon's maxCarrySize
field. Since that field gets used before applying
subskills, I'm not convinced I found the right
name, but I think the places that use it care more
about a specific Pokemon's carry limit than they
do about the species' theoretical max.