fix(test): cleanup ingredient-utils.test.ts
The 'combineSameIngredientsInDrop' test isn't something liable to change
when we change our prettify functions or when the game updates, so I
don't like using a snapshot test there. 'toStrictEqual' assets deep
equality of the arrays.

One of the other snapshot tests looks like it's saying the wrong thing.
This commit isn't fixing broken tests, so I'm adding a TODO for that
assertion.