chore(skill): Add test for stockpile
This skill tests that our assumption for Spit-Up-rate matches what we
expect to see: that the expected value of a skill trigger matches the
value of a 0-stack Spit Up. For the first 6 levels, this also matches
the RP value of the skill, similar to other (non-range) Charge Strength
skills.